requirements_val_api.txt
vllm==0.4.1
pydantic >= 2.0
transformers
torch
accelerate
bitsandbytes
pandas
sentencepiece
protobuf
python-dotenv
supabase
tiktoken
